soc/ace_v1x: Fix sched_ipi() delivery on core 0
The MP init path wasn't flagging the core active, so IPIs destined for
core 0 weren't being delivered from the code in the intel_adsp layer.
This led to some very strange behavior, in particular in ztest 1cpu
cases which would hang (presumably because core 0, if "held" during
the test, wouldn't wake back up at the end of a test case because the
IPI wouldn't arrive).

Signed-off-by: Enjia Mai <enjia.mai@intel.com>
Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>